pyproject.toml
[tool.poetry]
name = "wallet-api"
version = "0.0.0"
description = ""
authors = ["AATrubilin <aatrubilin@gmail.com>"]
readme = "README.md"
packages = [{include = "wallet_api"}]

[tool.poetry.dependencies]
django = "^5.0.2"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
djangorestframework-jsonapi = {extras = ["openapi"], version = "^6.1.0"}
mysqlclient = "^2.2.4"
python = "^3.10"
sentry-sdk = "^1.40.4"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
django-debug-toolbar = "^4.3.0"
isort = "^5.13.2"
pre-commit = "^3.6.1"
commitizen = "^3.14.1"


[tool.poetry.group.test.dependencies]
coverage = "^7.4.1"
django-coverage-plugin = "^3.1.0"
factory-boy = "^3.3.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.8.0"
pytest-sugar = "^1.0.0"


[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["wallet_api", "config"]
multi_line_output = 3
default_section = "THIRDPARTY"
skip = "venv/"
skip_glob = "**/migrations/*.py"
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true


[tool.pytest.ini_options]
addopts = "--ds=config.settings.test --reuse-db"
python_files = [
    "tests.py",
    "test_*.py"
]


[tool.coverage.run]
branch = true
source = ["wallet_api"]
omit = [
    "*/migrations/*",
]
plugins = ["django_coverage_plugin"]


[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version",
]
major_version_zero = true
annotated_tag = true


[tool.semantic_release]
version_toml = [
    "pyproject.toml:tool.poetry.version",
    "pyproject.toml:tool.commitizen.version",
]
version_variable = []
version_source = "commit"
major_on_zero = true
branch = "master"
hvcs = "github"
hvcs_domain = "github.com"
hvcs_api_domain = "github.com"
upload_to_release = true
upload_to_repository = true
build_command = "pip install poetry && poetry build"
commit_subject = "chore(release): {version}"
commit_message = "Automatically generated by python-semantic-release"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
